move test definitions to TS
